Override the UA string for facebook.com to ensure we’re getting touch-enabled content.
Without this override, we were getting unstyled mobile content from the 90s.
Impersonating an iphone or android would offer to install the respective applications when logging in.
The 'Firefox' token gets us the right content (and pretending to be AppleWebKit seems to be necessary too, otherwise the layout is busted).