Delay the auto-hide/auto-show behaviour of the header, in order to prevent the view from jumping up and down on touch-enabled devices when the touch event sequence is not finished.
Ideally, instead of a time-based delay, we should delay the auto-hide/auto-show behaviour until after the touch event sequence has finished (i.e. the finger was released), but to detect that in QML we would need to accept the first TouchBegin event, which would result in subsequent touch events not being forwarded to the webview underneath.