Style change...
First, thanks to @lduboeuf for this implementation! This is a great looking way to handle these menus, much easier to interact with than the tiny box that shows up in Demo Browser.

However I am curious why the dimming of the background was disabled? On a white page it makes the select dialogue look very undefined... I think at a minimum there should be a border to define the popup, or just letting the site be dimmed which I personally like.

Pictures to follow.