Create ActionList after the place that refers it
This is to workaround QObject children destruction order which destructs
in creation order. Otherwise, the ActionList and its children would be
destroyed before e.g. the Repeater, which will cause SIGSEGV at exit
when the Repeater is subsequently destroyed.

I don't understand why it's done this way; I would expect objects
declaratively created to be destroyed in the reverse order, in the same
way objects created on the stack in C++ would.