Set a default for offTheRecord in MorphWebContext
QtWebEngine 5.13 changed the default value of offTheRecord from false to
true. This means that apps will dump all cookies and site data on exit
by default.

If someone is using the MorphWebContext as part of their app, they
expect the old behavior with offTheRecord = false. So we'll just set
that.

https://github.com/ubports/morph-browser/issues/347

This does not help apps which are using WebEngineProfile directly
without setting offTheRecord. Those apps will need to set
"offTheRecord: false" on their Profile by themselves.