Adds navigation history in the back/forward button (#444)
Press-and-hold and right-click on the back and forward buttons will open a popup with the navigation history just like on other browsers. Currently, it uses Popover with custom background color since it doesn't "popup" enough on white contents which websites usually are. I will try to use Popup/Menu from QQC2 and see if it looks better.

This is an expansion of the initial effort in #428.